# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Mengambil daftar employee",
                "operationId": "employeesIndex",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter berdasarkan name, nik_telin, email, atau status_login",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar employee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Membuat employee baru",
                "operationId": "employeesStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/employees/{emp_id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Display the specified resource.",
                "description": "Mengambil detail employee",
                "operationId": "employeesShow",
                "parameters": [
                    {
                        "name": "emp_id",
                        "in": "path",
                        "description": "ID employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail employee",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Memperbarui data employee",
                "operationId": "employeesUpdate",
                "parameters": [
                    {
                        "name": "emp_id",
                        "in": "path",
                        "description": "ID employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "Menghapus employee",
                "operationId": "employeesDestroy",
                "parameters": [
                    {
                        "name": "emp_id",
                        "in": "path",
                        "description": "ID employee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Employee berhasil dihapus"
                    },
                    "404": {
                        "description": "Employee tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Employee": {
                "description": "Model Employee",
                "properties": {
                    "emp_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "nik_telin": {
                        "type": "string",
                        "nullable": true
                    },
                    "nik_tg": {
                        "type": "string",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "nullable": true
                    },
                    "religion": {
                        "type": "string",
                        "nullable": true
                    },
                    "place_of_birth": {
                        "type": "string",
                        "nullable": true
                    },
                    "date_of_birth": {
                        "type": "string",
                        "nullable": true
                    },
                    "employee_status": {
                        "type": "string",
                        "nullable": true
                    },
                    "position_band": {
                        "type": "string",
                        "nullable": true
                    },
                    "work_location": {
                        "type": "string",
                        "nullable": true
                    },
                    "office_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "type": "string",
                        "nullable": true
                    },
                    "province": {
                        "type": "string",
                        "nullable": true
                    },
                    "pos_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "telepon": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_city": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_province": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_country": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_pos_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "my_tlp_no": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "email_asli": {
                        "type": "string",
                        "nullable": true
                    },
                    "profile_photo": {
                        "type": "string",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_status_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "auth_type_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "auth_dtm": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "nonce": {
                        "type": "string",
                        "nullable": true
                    },
                    "token": {
                        "type": "string",
                        "nullable": true
                    },
                    "attempts": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status_login": {
                        "type": "string",
                        "nullable": true
                    },
                    "positions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Position"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "EmployeeRequest": {
                "description": "Payload untuk membuat/memperbarui Employee",
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "nik_telin": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "nik_tg": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "gender": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "religion": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "place_of_birth": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "date_of_birth": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "employee_status": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "position_band": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "work_location": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "office_address": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 200
                    },
                    "city": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "province": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "pos_code": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "telepon": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "my_address": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 200
                    },
                    "my_city": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "my_province": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "my_country": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "my_pos_code": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "my_tlp_no": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "email": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "email_asli": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "profile_photo": {
                        "type": "string",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 120
                    },
                    "language": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    },
                    "user_status_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "auth_type_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "auth_dtm": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "nonce": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 32
                    },
                    "token": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 32
                    },
                    "attempts": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status_login": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 45
                    }
                },
                "type": "object"
            },
            "Position": {
                "description": "Model Position",
                "properties": {
                    "pos_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "department": {
                        "type": "object",
                        "nullable": true
                    },
                    "pivot": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}